fix: flaky test with `go 1.20` (#302)
* add codeowners

* add owners

* fix flaky test

* Revert "add codeowners"

This reverts commit 17274055

* Quo

+++ x/interchainstaking/keeper/ibc_packet_handlers_test.go
@@ -777,7 +777,9 @@ func (s *KeeperTestSuite) TestReceiveAckErrForBeginUndelegate() {
 	hash3 := fmt.Sprintf("%x", sha256.Sum256([]byte{0x03}))
 	delegator1 := utils.GenerateAccAddressForTest().String()
 	delegator2 := utils.GenerateAccAddressForTest().String()
-	randRr := rand.Float64() + 1.0
+
+	randRr := rand.Float32() + 1.0
+
 	tests := []struct {
 		name                      string
 		epoch                     int64
